@@ -88,15 +88,15 @@ public class ReporterPlayerListener implements Listener {
 	private void listOnLogin(final Player player) {
 		final ReporterCommand listCommand = plugin.getCommandManager().getCommand(ListCommand.getCommandName());
 		if (listCommand.hasPermission(player)) {
-			listCommand.setSender(player);
-			listCommand.setArguments(new ArrayList<String>());
-			if (plugin.getConfig().getBoolean("general.messaging.listOnLogin.useDelay", true)) {
-				final int delay = plugin.getConfig().getInt("general.messaging.listOnLogin.delay", 5);
-				Bukkit.getScheduler().runTaskLaterAsynchronously(plugin, listCommand, BukkitUtil.convertSecondsToServerTicks(delay));
-			} else {
-				Bukkit.getScheduler().runTaskAsynchronously(plugin, listCommand);
-			}
-		}
+            listCommand.setSender(player);
+            listCommand.setArguments(new ArrayList<>());
+            if (plugin.getConfig().getBoolean("general.messaging.listOnLogin.useDelay", true)) {
+                final int delay = plugin.getConfig().getInt("general.messaging.listOnLogin.delay", 5);
+                Bukkit.getScheduler().runTaskLaterAsynchronously(plugin, listCommand, BukkitUtil.convertSecondsToServerTicks(delay));
+            } else {
+                Bukkit.getScheduler().runTaskAsynchronously(plugin, listCommand);
+            }
+        }
 	}
 
 	private void sendMessages(final Player player) {
