@@ -62,67 +62,67 @@ public class ModLevelTest {
 
 	@Test
 	public void testCompareToByLevelModLevelModLevel() {
-		assertTrue(ModLevel.compareToByLevel(ModLevel.UNKNOWN, ModLevel.UNKNOWN) == 0);
+		assertEquals(0, ModLevel.compareToByLevel(ModLevel.UNKNOWN, ModLevel.UNKNOWN));
 		assertTrue(ModLevel.compareToByLevel(ModLevel.UNKNOWN, ModLevel.NONE) < 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.UNKNOWN, ModLevel.LOW) < 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.UNKNOWN, ModLevel.NORMAL) < 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.UNKNOWN, ModLevel.HIGH) < 0);
 
 		assertTrue(ModLevel.compareToByLevel(ModLevel.NONE, ModLevel.UNKNOWN) > 0);
-		assertTrue(ModLevel.compareToByLevel(ModLevel.NONE, ModLevel.NONE) == 0);
+		assertEquals(0, ModLevel.compareToByLevel(ModLevel.NONE, ModLevel.NONE));
 		assertTrue(ModLevel.compareToByLevel(ModLevel.NONE, ModLevel.LOW) < 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.NONE, ModLevel.NORMAL) < 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.NONE, ModLevel.HIGH) < 0);
 
 		assertTrue(ModLevel.compareToByLevel(ModLevel.LOW, ModLevel.UNKNOWN) > 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.LOW, ModLevel.NONE) > 0);
-		assertTrue(ModLevel.compareToByLevel(ModLevel.LOW, ModLevel.LOW) == 0);
+		assertEquals(0, ModLevel.compareToByLevel(ModLevel.LOW, ModLevel.LOW));
 		assertTrue(ModLevel.compareToByLevel(ModLevel.LOW, ModLevel.NORMAL) < 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.LOW, ModLevel.HIGH) < 0);
 
 		assertTrue(ModLevel.compareToByLevel(ModLevel.NORMAL, ModLevel.UNKNOWN) > 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.NORMAL, ModLevel.NONE) > 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.NORMAL, ModLevel.LOW) > 0);
-		assertTrue(ModLevel.compareToByLevel(ModLevel.NORMAL, ModLevel.NORMAL) == 0);
+		assertEquals(0, ModLevel.compareToByLevel(ModLevel.NORMAL, ModLevel.NORMAL));
 		assertTrue(ModLevel.compareToByLevel(ModLevel.NORMAL, ModLevel.HIGH) < 0);
 
 		assertTrue(ModLevel.compareToByLevel(ModLevel.HIGH, ModLevel.UNKNOWN) > 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.HIGH, ModLevel.NONE) > 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.HIGH, ModLevel.LOW) > 0);
 		assertTrue(ModLevel.compareToByLevel(ModLevel.HIGH, ModLevel.NORMAL) > 0);
-		assertTrue(ModLevel.compareToByLevel(ModLevel.HIGH, ModLevel.HIGH) == 0);
+		assertEquals(0, ModLevel.compareToByLevel(ModLevel.HIGH, ModLevel.HIGH));
 	}
 
 	@Test
 	public void testCompareToByLevelModLevel() {
-		assertTrue(ModLevel.UNKNOWN.compareToByLevel(ModLevel.UNKNOWN) == 0);
+		assertEquals(0, ModLevel.UNKNOWN.compareToByLevel(ModLevel.UNKNOWN));
 		assertTrue(ModLevel.UNKNOWN.compareToByLevel(ModLevel.NONE) < 0);
 		assertTrue(ModLevel.UNKNOWN.compareToByLevel(ModLevel.LOW) < 0);
 		assertTrue(ModLevel.UNKNOWN.compareToByLevel(ModLevel.NORMAL) < 0);
 		assertTrue(ModLevel.UNKNOWN.compareToByLevel(ModLevel.HIGH) < 0);
 
 		assertTrue(ModLevel.NONE.compareToByLevel(ModLevel.UNKNOWN) > 0);
-		assertTrue(ModLevel.NONE.compareToByLevel(ModLevel.NONE) == 0);
+		assertEquals(0, ModLevel.NONE.compareToByLevel(ModLevel.NONE));
 		assertTrue(ModLevel.NONE.compareToByLevel(ModLevel.LOW) < 0);
 		assertTrue(ModLevel.NONE.compareToByLevel(ModLevel.NORMAL) < 0);
 		assertTrue(ModLevel.NONE.compareToByLevel(ModLevel.HIGH) < 0);
 
 		assertTrue(ModLevel.LOW.compareToByLevel(ModLevel.UNKNOWN) > 0);
 		assertTrue(ModLevel.LOW.compareToByLevel(ModLevel.NONE) > 0);
-		assertTrue(ModLevel.LOW.compareToByLevel(ModLevel.LOW) == 0);
+		assertEquals(0, ModLevel.LOW.compareToByLevel(ModLevel.LOW));
 		assertTrue(ModLevel.LOW.compareToByLevel(ModLevel.NORMAL) < 0);
 		assertTrue(ModLevel.LOW.compareToByLevel(ModLevel.HIGH) < 0);
 
 		assertTrue(ModLevel.NORMAL.compareToByLevel(ModLevel.UNKNOWN) > 0);
 		assertTrue(ModLevel.NORMAL.compareToByLevel(ModLevel.NONE) > 0);
 		assertTrue(ModLevel.NORMAL.compareToByLevel(ModLevel.LOW) > 0);
-		assertTrue(ModLevel.NORMAL.compareToByLevel(ModLevel.NORMAL) == 0);
+		assertEquals(0, ModLevel.NORMAL.compareToByLevel(ModLevel.NORMAL));
 		assertTrue(ModLevel.NORMAL.compareToByLevel(ModLevel.HIGH) < 0);
 
 		assertTrue(ModLevel.HIGH.compareToByLevel(ModLevel.UNKNOWN) > 0);
 		assertTrue(ModLevel.HIGH.compareToByLevel(ModLevel.NONE) > 0);
 		assertTrue(ModLevel.HIGH.compareToByLevel(ModLevel.LOW) > 0);
 		assertTrue(ModLevel.HIGH.compareToByLevel(ModLevel.NORMAL) > 0);
-		assertTrue(ModLevel.HIGH.compareToByLevel(ModLevel.HIGH) == 0);
+		assertEquals(0, ModLevel.HIGH.compareToByLevel(ModLevel.HIGH));
 	}
 }
